requirements.txt
PyQt6
google-genai
pdfplumber
python-docx
reportlab
Pillow
